Finished build scripts and fixed random weather generation algorithm to include random biomes

+++ build.sh
@@ -1,5 +1,9 @@
-rm -rf build
-find . -name '*.java' > sources.txt
+[ -d build ] && rm -rf build
+find . -name *.java > result.txt
+sort result.txt > sources.txt
+mkdir build
 javac src/com/avaj/Main.java -sourcepath @sources.txt -d build
+rm -rf result.txt
 rm -rf sources.txt
+
